+++ backEnd/learnkit-backend/src/main/java/com/learnkit/backend/exception/custom/GoalNotFoundException.java
@@ -1,7 +1,7 @@
 package com.learnkit.backend.exception.custom;
 
 public class GoalNotFoundException extends RuntimeException{
-    public GoalNotFoundException(Long goalId) {
+    public GoalNotFoundException(Integer goalId) {
         super("Goal not found with id: " + goalId);
     }
 }
